requirements.txt
absolufy-imports==0.3.1
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
Authlib==1.6.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
cryptography==45.0.7
docstring_parser==0.17.0
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.9.0
google-adk==1.13.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.111.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.37.0
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.57.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.35.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
hpack==4.1.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.2
mcp==1.13.1
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
opentelemetry-api==1.36.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.2
qdrant-client==1.15.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.47.3
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
websockets==15.0.1
zipp==3.23.0
